Squashed commit of the following:
commit 5e36b170a14a16003033d8d0edaf15230cd8ceb0
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sun Aug 12 22:10:07 2018 +0300

    grand squash commit

commit 02a99dd15deaecda14efe7246cad2043da063759
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat Jun 2 13:47:26 2018 +0300

    pause method added

commit 4ab8521f86a901db76de0de4d4c176ddd2692c69
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat Jun 2 13:39:48 2018 +0300

    fx_looper fixes

commit 68e2a58958386524c5371e922d383c2fa01b9c0c
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Thu May 31 15:36:55 2018 +0300

    looper faders added

commit e4fb4fc8e5d6c76c839be00b43eaba564a3b1df1
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat May 26 15:05:31 2018 +0300

    list.do data support

commit deb79f893ca752c1f70bdafede03a3a5228c78c1
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat May 26 14:21:37 2018 +0300

    list.count_if data support

commit 1262d4e4492e72a50fdc4e9a8404c968b7b0d95c
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat May 26 13:36:58 2018 +0300

    list.count data support

commit 77173f1c6f870fa2baeab452f0a47c44afe37826
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Thu May 24 02:48:01 2018 +0300

    doc update

commit 9951de52106af7c0fb4d81bde731e3fc4d480a15
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Thu May 24 02:45:19 2018 +0300

    list.at mlist support added

commit 1094a3fb57d7933d3ecdbb467debb7465beca59a
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Thu May 24 02:29:16 2018 +0300

    list.all_off  update

commit a781686f3df984e31dd28337d2e98cb31d1090de
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Tue May 22 03:24:08 2018 +0300

    some docs added

commit cfe314a44a0d2cfe7d2fa24b6a60dfebc114e710
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Tue May 22 02:40:23 2018 +0300

    list.all_of mlist support

commit f9d4fe774671f9d2f4b762a6c64bbde83e5337d7
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Tue May 22 02:35:38 2018 +0300

    list.choice support added

commit 5b4e215fd8c75ccac07dadc391f1ac2b8d991533
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 21 01:51:37 2018 +0300

    remove method added

commit e10cbeece15648a9e93f8f5eb9d2ac2d06abe732
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 21 01:40:39 2018 +0300

    pop method added

commit 07c6693a59dd68ef3ea96df7b3d5d6413f74e9b3
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Thu May 17 03:14:55 2018 +0300

    mlist update

commit 7f578ea0344e8e6f3a98fcfa5efc90adc7742c30
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Thu May 17 03:14:45 2018 +0300

    tests update

commit bfa088f0bff81aa6ec064a5ac3e4fee3b5baa1d7
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Thu May 17 00:38:56 2018 +0300

    insert method added

commit 898d97d50e1a31a7e48ea1335470b947f954b890
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Wed May 16 03:19:45 2018 +0300

    global.mlist added

commit 909e8184887730864f5cffe05bb7fe58fbdb2e62
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Wed May 16 03:11:45 2018 +0300

    insert method added

commit fbe5154b3f30f45589981ea4de4da64f2953f452
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Wed May 16 02:50:46 2018 +0300

    clear method added

commit 69c4867b3a2724fa8f245b7e59e4c401213ef1bb
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Wed May 16 01:46:39 2018 +0300

    new data functions added

commit 65049b8507c8aa43bc88fe195a0116b9b66c0435
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Wed May 16 00:47:41 2018 +0300

    data support added to Message class

commit ad863bbd4b8876dcdb19ae9c75d4da3d1f1db6db
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Wed May 16 00:30:42 2018 +0300

    'at' method added

commit 7f2faa82eec9b4d4765646410c6dcb38146affa9
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 17:44:42 2018 +0300

    more tests

commit 32facddbc9ab6ff4b6d13ed0f9689ed011e12a7c
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 16:44:30 2018 +0300

    single quote replaced with `

    To escape ` use '`

    The problem is in Pd GUI you can't enter \\ (backslash) characted, so to avoid hard escaping  less  common symbol used for strings then single quote

commit 3f0e37a83e3274e8a579c8e90a14bd06e86c427d
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 16:14:55 2018 +0300

    base data.mlist added

commit cbd2d1acef8f344a38eb49b69c2754ae4dac2a0c
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 15:44:34 2018 +0300

    migration to data directory

commit 2eb9f45d1823c2e6e1be3dd0187539c8a1f14d14
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 15:33:40 2018 +0300

    more tests

commit 8ea535c3f3a3e72bffce793ad550436c865b6ab3
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 15:28:28 2018 +0300

    mlist fixes

commit a5e5ed7248c7817a32c9badb42471a65279995c3
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 15:05:52 2018 +0300

    quotes formatting added

commit 51c81436f81bc0d427b915216c9d6913667c3498
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 14:03:17 2018 +0300

    mlist data

commit d9004d57780618519ad833fedb296c01970f41c7
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 12:47:23 2018 +0300

    ast traverse added

commit 2e63c461f24f94d306e4632afe6671b684a18769
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Mon May 14 12:24:22 2018 +0300

    test fix

commit 358ebf3157d0074bf8b586ff21a3666abb9b9baa
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sun May 13 11:18:39 2018 +0300

    more tests

commit 6e0feec40321dd0542a778314d30c7b73e1e114e
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sun May 13 01:11:43 2018 +0300

    more cleanup

commit 910e33bb405586fe3bb4057404554470c4bbfbd6
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sun May 13 01:06:24 2018 +0300

    cleanup

commit 6b642714d25a9e0d5c83bda8fa4c5c076938b9df
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sun May 13 01:03:43 2018 +0300

    first working example

commit d331d300b3ec81c31ad295d2a1d037b01ccbbd98
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat May 12 02:21:33 2018 +0300

    impl functions

commit 2ffb198698af0e5ab1370dada6008e1148a4e50e
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat May 12 02:13:46 2018 +0300

    impl hide

commit 2edec93471c079386391b6ce4292804eb4d48494
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat May 12 02:05:05 2018 +0300

    struct added

commit f9b2d9f700942aea5824814956d667941d84f77c
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat May 12 01:52:38 2018 +0300

    mote tests

commit 6d1fff6e696e325f36d7037c109fa08dc64eed16
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat May 12 01:30:09 2018 +0300

    working example

commit e6f81e70867ff9828b19e35ffb69c03e84a7c606
Author: Serge Poltavski <serge.poltavski@gmail.com>
Date:   Sat May 12 00:35:33 2018 +0300

    compile state fixing

+++ ceammc/ext/src/list/list_seq.cpp
@@ -1,10 +1,8 @@
 #include <algorithm>
+#include <cmath>
 
 #include "ceammc_factory.h"
 #include "list_seq.h"
-
-#include "ceammc.hpp"
-#include <m_pd.h>
 
 ListSeq::ListSeq(const PdArgs& a)
     : BaseObject(a)
@@ -47,7 +45,7 @@ void ListSeq::onBang()
     AtomList res;
     const t_float from = from_->value();
     const t_float to = to_->value();
-    const t_float step = fabsf(step_->value());
+    const t_float step = std::fabs(step_->value());
 
     if (step == 0.f) {
         OBJ_ERR << "invalid step: " << step;
